incorporate suggestions
